// src/product-purchase/interface/productPurchase.interface.ts
export interface ProductPuchaseInterface {
  prodId: String
  purchaseOrderNo: String
  paymentStatus: String
  amountPaid: Number
  expectedAt: Date
  deliveredAt: Date
  GST: number
  Rate_Unit: Number
  GSTValue: Number
  totalValue: Number
  costPrice: number
  quantity: Number
  manufacturer: String
  manufacturerId: String
  closedBy: String
  closedById: String
  closedByRole: String
  closedAt: Date
  paidBy: String
  paidById: String
  paidByRole: String
  paidAt: Date
};